% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{data_processing}
\alias{data_processing}
\title{sellke}
\usage{
data_processing(n, rho, beta, gamma, T.max_analysis = 10, nu = NULL)
}
\arguments{
\item{n}{Interger. Number of initial susceptible.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{nu}{Numeric. A parameter denoting the standard deviation of the frailty variable which is assumed to follow Gamma distribution.}

\item{Tmax}{Numeric. Final observation time of epidemic.}
}
\value{
A data frame with two columns - days and daily infection counts respectively.
}
\description{
Generates a dataset using the Sellke construction and then produce a data frame with two columns: days and incidence counts. If \code{nu} is specified, it simulate trajectories incorporating frailty variable.
}
