% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSA_data.R
\name{DSA_data}
\alias{DSA_data}
\title{DSA_data}
\usage{
DSA_data(n, beta, gamma, rho, Tmax, dt = 0.01)
}
\arguments{
\item{n}{Interger. Number of initial susceptible.}

\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic}

\item{dt}{Numeric. time increment}
}
\value{
A data set with two columns - exact infection times and recovery times respectively
}
\description{
Generates a data set of exact infection and recovery times according to DSA approximation for simulating epidemic trajectories.
}
