% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_tau.R
\name{f_tau}
\alias{f_tau}
\title{f_tau}
\usage{
f_tau(beta, gamma, rho, Tmax, dt = 0.1)
}
\arguments{
\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic}

\item{dt}{Numeric. Time increment.}
}
\value{
A matrix with time points from zero to final time, and its corresponding value of the density  under standard SIR model at every time point.
}
\description{
Assuming standard SIR model, the function calculates the value of the density of infection times for all values from zero to Final observation time given the parameters.
}
