% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_tau_fr.R
\name{f_tau_fr}
\alias{f_tau_fr}
\title{f_tau_fr}
\usage{
f_tau_fr(beta, gamma, rho, nu, Tmax, dt = 0.1)
}
\arguments{
\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < \eqn{\rho}{rho} < 1.}

\item{nu}{Numeric. A parameter denoting the standard deviation of the frailty variable which is assumed to follow Gamma distribution.}

\item{Tmax}{Numeric. Final observation time of epidemic.}

\item{dt}{Numeric. time increment}
}
\value{
A matrix with time points from zero to final time, and its corresponding value of the density under frailty model.
}
\description{
When we assume a frailty model, the function calculates the value of the density of infection times for all values from zero to Final observation time given the value of the parameters.
}
