% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_lkd.R
\name{full_lkd}
\alias{full_lkd}
\title{full_lkd}
\usage{
full_lkd(
  data,
  T.max_analysis,
  iteration = 10000,
  num_chain = 4,
  num_cores = 1,
  stan_seed = 12,
  stan_warmup = 0
)
}
\arguments{
\item{data}{Matrix/Data Frame with two column continuous infection and recovery times respectively.}

\item{T.max_analysis}{Numeric. Final observation time of epidemic}

\item{iteration}{Integer. The number of iterations for each chain (including warmup). The default is 10000.}

\item{num_chain}{Integer specifying the number of Markov chains. The default is 4.}

\item{num_cores}{Same as "cores" in Sampling function of Stan. Default is 1.}

\item{stan_seed}{Same as "seed" in Sampling function of Stan. Default is 12.}

\item{stan_warmup}{Same as "warmup" in Sampling function of Stan. Default is floor(iteration/2).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling()},
containing the posterior samples for the parameters \code{beta}, \code{gamma}, \code{rho}, and \code{R_0}.
}
\description{
Taking a matrix or data frame with exact infection times, recovery times respectively in two columns, the function do inference of parameters using complete likelihood in Stan.
}
