% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_frailty_ODE.R
\name{limiting_frailty_ODE}
\alias{limiting_frailty_ODE}
\title{limiting_frailty_ODE}
\usage{
limiting_frailty_ODE(beta, gamma, rho, nu, Tmax, dt = 0.01)
}
\arguments{
\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < \eqn{\rho}{rho} < 1.}

\item{nu}{Numeric. A parameter denoting the standard deviation of the frailty variable which is assumed to follow Gamma distribution.}

\item{Tmax}{Numeric. Final observation time of epidemic.}

\item{dt}{Numeric. time increment}
}
\value{
Returns the limiting ODEs of frailty model.
}
\description{
Assuming frailty model, given the parameters, calculates the value of the limiting ODEs.
}
