% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_SIR_ODE.R
\name{limiting_SIR_ODE}
\alias{limiting_SIR_ODE}
\title{limiting_SIR_ODE}
\usage{
limiting_SIR_ODE(beta, gamma, rho, Tmax, dt = 0.01)
}
\arguments{
\item{beta}{Numeric. Infection rate parameter.}

\item{gamma}{Numeric. Recovery rate parameter.}

\item{rho}{Numeric. Limiting proportion of initially infected individuals relative to the initially susceptible population. Must satisfy 0 < rho < 1.}

\item{Tmax}{Numeric. Final observation time of epidemic}

\item{dt}{Numeric. time increment}
}
\value{
Returns the limiting SIR ODEs.
}
\description{
Assuming standard SIR model, given the parameters, calculates the value of the limiting ODEs.
}
