% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_ct_lkd.R
\name{new_data_ct_lkd}
\alias{new_data_ct_lkd}
\title{new_data_ct_lkd}
\usage{
new_data_ct_lkd(
  time_points,
  infection_count,
  Final_time,
  initial_sus = 0,
  frailty = FALSE,
  iteration = 10000,
  num_chain = 4,
  num_cores = 1,
  stan_seed = 12,
  stan_warmup = 0
)
}
\arguments{
\item{time_points}{A numeric vector giving the right endpoints of the time intervals into which the observation period is divided. For example, if intervals are \((0,1], (1,2], \dots\), then \code{time_points = c(1, 2, ...)}. It is assumed that intervals are of equal length and starts from 0.}

\item{infection_count}{A vector of integers giving the infection counts. We suggest to keep the time points equally spaced between one to the final time and adjust the infection count vector such that it takes value zero, if we do not observe infections in some of the days.}

\item{Final_time}{Numeric. Final observation time of epidemic.}

\item{initial_sus}{If available, give the initial number of susceptible, otherwise, put 0.}

\item{frailty}{Logical. If True, makes inference using frailty model and standard SIR model otherwise.}

\item{iteration}{Integer. The number of iterations for each chain (including warmup). The default is 10000.}

\item{num_chain}{Integer specifying the number of Markov chains. The default is 4.}

\item{num_cores}{Same as "cores" in Sampling function of Stan. Default is 1.}

\item{stan_seed}{Same as "seed" in Sampling function of Stan. Default is 12.}

\item{stan_warmup}{Same as "warmup" in Sampling function of Stan. Default is floor(iteration/2).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling()},
containing the posterior samples for the parameters \eqn{\beta}{beta}, \eqn{\gamma}{gamma}, \eqn{\rho}{rho}.
}
\description{
The function returns a Stan object containing the posterior samples for the parameters beta, gamma, rho ( and N if not available).
}
